Update and rename Final to TestTubePickerFinal
Here the robot does the homing once it is connected to the power source
We can control the robot manually in this case. We have the 6 containers labeled as A, B, C, D, E, and F. 
The gripper can be opened by pressing O for opening and P for closing.